fixed projectile spawn and slow wiggle pattern
