+++ Roche-Engine/Resources/Entity/Entity_Game.json
@@ -36,7 +36,7 @@
             45.0
         ],
         "projectileBullet": "Projectile1",
-        "projectilePattern": "Resources\\Patterns\\DefaultPattern.ptn",
+        "projectilePattern": "Resources\\Patterns\\SlowWiggle.ptn",
         "projectileSystem": true,
         "rotation": 0.0,
         "rows": 4,
